# docs/swagger.yaml
definitions:
...
  LikeRequest:
    properties:
      album_id:
        type: integer
      photo_id:
        type: integer
    type: object
...
  Photo:
    properties:
      album_id:
        type: integer
      commentCount:
        type: integer
      created_at:
        type: string
      deleted_at:
        type: string
      id:
        type: integer
      likeCount:
        type: integer
      url_image:
        type: string
    type: object
...
  User:
    properties:
      biography:
        type: string
      birth:
        type: string
      created_at:
        type: string
      ddd:
        type: string
      email:
        type: string
      first_name:
        type: string
      id:
        type: integer
      last_name:
        type: string
      password:
        type: string
      phone:
        type: string
      updated_at:
        type: string
      urlImage:
        type: string
      username:
        type: string
    type: object
  UserCreateRequest:
    properties:
      email:
        type: string
      first_name:
        type: string
      last_name:
        type: string
      password:
        type: string
      username:
        type: string
    type: object
  UserCreateResponse:
    properties:
      email:
        type: string
      first_name:
        type: string
      id:
        type: integer
      last_name:
        type: string
      username:
        type: string
    type: object
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
paths:
...
  /like:
    post:
      consumes:
      - application/json
      description: '- give a likes'
      parameters:
      - description: Like to be created
        in: body
        name: Like
        required: true
        schema:
          $ref: '#/definitions/LikeRequest'
      produces:
      - application/json
      responses:
        "201":
          description: ""
      security:
      - ApiKeyAuth: []
      summary: '- give a likes'
      tags:
      - '- Like'
  /like/{album_id}/album:
    get:
      consumes:
      - application/json
      description: '- show all likes of the albums'
      parameters:
      - description: album_id
        in: path
        name: album_id
        required: true
        type: integer
      - description: limit per page
        in: query
        name: limit
        required: true
        type: integer
      - description: number of page
        in: query
        name: page
        required: true
        type: integer
      - description: search sort
        in: query
        name: sort
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: ""
      security:
      - ApiKeyAuth: []
      summary: '- show all likes of the albums'
      tags:
      - '- Like'
  /like/{id}:
    delete:
      consumes:
      - application/json
      description: '- delete a likes'
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: ""
      security:
      - ApiKeyAuth: []
      summary: '- delete a likes'
      tags:
      - '- Like'
  /like/photo/{photo_id}:
    get:
      consumes:
      - application/json
      description: '- show all likes of the photos'
      parameters:
      - description: photo_id
        in: path
        name: photo_id
        required: true
        type: integer
      - description: limit per page
        in: query
        name: limit
        required: true
        type: integer
      - description: number of page
        in: query
        name: page
        required: true
        type: integer
      - description: search sort
        in: query
        name: sort
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: ""
      security:
      - ApiKeyAuth: []
      summary: '- show all likes of the photos'
      tags:
      - '- Like'
  /photo/{album_id}:
    get:
      consumes:
      - application/json
      description: '- get all photos by album'
      parameters:
      - description: album_id
        in: path
        name: album_id
        required: true
        type: integer
      - description: limit per page
        in: query
        name: limit
        required: true
        type: integer
      - description: number of page
        in: query
        name: page
        required: true
        type: integer
      - description: search sort
        in: query
        name: sort
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/Photo'
            type: array
      security:
      - ApiKeyAuth: []
      summary: '- get all photos'
      tags:
      - '- Photo'
    post:
      consumes:
      - multipart/form-data
      description: '- Upload a photo'
      parameters:
      - description: album_id
        in: path
        name: album_id
        required: true
        type: integer
      - description: upload your file
        in: formData
        name: file
        required: true
        type: file
      produces:
      - application/json
      responses:
        "204":
          description: ""
      security:
      - ApiKeyAuth: []
      summary: '- Upload photo'
      tags:
      - '- Photo'
  /photo/{photo_id}:
    delete:
      consumes:
      - application/json
      description: '- delete a photo'
      parameters:
      - description: photo_id
        in: path
        name: photo_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: ""
      security:
      - ApiKeyAuth: []
      summary: '- delete a photo'
      tags:
      - '- Photo'
  /photo/album/{photo_id}:
    get:
      consumes:
      - application/json
      description: '- get photo by id'
      parameters:
      - description: photo_id
        in: path
        name: photo_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/Photo'
      security:
      - ApiKeyAuth: []
      summary: '- get photo'
      tags:
      - '- Photo'
  /user:
    post:
      consumes:
      - application/json
      description: '- Create a user'
      parameters:
      - description: User to be created
        in: body
        name: User
        required: true
        schema:
          $ref: '#/definitions/UserCreateRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/UserCreateResponse'
      security:
      - ApiKeyAuth: []
      summary: '- Create user'
      tags:
      - '- User'
  /user/{id}:
    delete:
      consumes:
      - application/json
      description: '- Delete a user by id'
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: ""
      security:
      - ApiKeyAuth: []
      summary: '- Delete user'
      tags:
      - '- User'
    get:
      consumes:
      - application/json
      description: Get user by id
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/User'
      security:
      - ApiKeyAuth: []
      summary: '- Get one user'
      tags:
      - '- User'
    put:
      consumes:
      - application/json
      description: '- Update a user by id'
      parameters:
      - description: User to be updated
        in: body
        name: User
        required: true
        schema:
          $ref: '#/definitions/User'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/User'
      security:
      - ApiKeyAuth: []
      summary: '- Update user'
      tags:
      - '- User'
  /user/upload/{id}:
    post:
      consumes:
      - application/json
      description: '- upload a profile picture'
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: integer
      - description: upload your file
        in: formData
        name: file
        required: true
        type: file
      produces:
      - application/json
      responses:
        "204":
          description: ""
      security:
      - ApiKeyAuth: []
      summary: '- upload a profile picture'
      tags:
      - '- User'
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"